refactor: extract functions for sanitizing torrent description
Decouple function from Vue component so that:

- They can be tested independently
- They can be reused easily in other components
